# backend-java/agent-api/src/main/resources/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Agent Ops API",
    "description": "Agent Ops 后端接口文档，可直接导入 Apifox 管理接口契约。",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地 Java 后端"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/kb": {
      "get": {
        "tags": ["ExternalKnowledge"],
        "summary": "查询知识库列表",
        "responses": {
          "200": { "description": "知识库列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/KnowledgeBaseResponse" } } } } }
        }
      },
      "post": {
        "tags": ["ExternalKnowledge"],
        "summary": "创建知识库",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeBaseRequest" } } } },
        "responses": {
          "200": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdResponse" } } } }
        }
      }
    },
    "/api/kb/{id}": {
      "get": {
        "tags": ["ExternalKnowledge"],
        "summary": "查询知识库详情",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "知识库详情", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeBaseResponse" } } } }
        }
      },
      "put": {
        "tags": ["ExternalKnowledge"],
        "summary": "更新知识库",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeBaseUpdateRequest" } } } },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["ExternalKnowledge"],
        "summary": "删除知识库",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "删除成功" } }
      }
    },
    "/api/kb/documents": {
      "post": {
        "tags": ["ExternalKnowledge"],
        "summary": "登记文档元数据",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentCreateRequest" } } } },
        "responses": {
          "200": { "description": "登记成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdResponse" } } } }
        }
      }
    },
    "/api/kb/{kbId}/documents/upload": {
      "post": {
        "tags": ["ExternalKnowledge"],
        "summary": "上传文档并触发入库",
        "parameters": [{ "$ref": "#/components/parameters/KbId" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "title": { "type": "string", "description": "文档标题；为空时使用文件名" },
                  "file": { "type": "string", "format": "binary", "description": "Markdown 或 TXT 文件" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "上传成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeDocumentResponse" } } } }
        }
      }
    },
    "/api/kb/{kbId}/documents": {
      "get": {
        "tags": ["ExternalKnowledge"],
        "summary": "查询知识库文档列表",
        "parameters": [{ "$ref": "#/components/parameters/KbId" }],
        "responses": {
          "200": { "description": "文档列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/KnowledgeDocumentResponse" } } } } }
        }
      }
    },
    "/api/kb/documents/{id}": {
      "get": {
        "tags": ["ExternalKnowledge"],
        "summary": "查询文档详情",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "文档详情", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeDocumentResponse" } } } }
        }
      },
      "put": {
        "tags": ["ExternalKnowledge"],
        "summary": "更新文档元数据",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentUpdateRequest" } } } },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["ExternalKnowledge"],
        "summary": "删除文档",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "删除成功" } }
      }
    },
    "/api/kb/documents/{id}/retry": {
      "post": {
        "tags": ["ExternalKnowledge"],
        "summary": "重试文档入库",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "重试已触发", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KnowledgeDocumentResponse" } } } }
        }
      }
    },
    "/api/kb/documents/{id}/download": {
      "get": {
        "tags": ["ExternalKnowledge"],
        "summary": "下载文档原文件",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "文件流", "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
      "KbId": { "name": "kbId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
...
    },
...
    "schemas": {
...
      "KnowledgeBaseRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PRIVATE", "TEAM", "PUBLIC"], "default": "PRIVATE" }
        }
      },
      "KnowledgeBaseUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PRIVATE", "TEAM", "PUBLIC"] }
        }
      },
      "KnowledgeBaseResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "visibility": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "DocumentCreateRequest": {
        "type": "object",
        "required": ["kbId", "title"],
        "properties": {
          "kbId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "sourceType": { "type": "string", "default": "markdown" },
          "sourceUri": { "type": "string" }
        }
      },
      "DocumentUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "sourceType": { "type": "string" },
          "sourceUri": { "type": "string" }
        }
      },
...
      "KnowledgeDocumentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "kbId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "sourceType": { "type": "string" },
          "sourceUri": { "type": "string" },
          "version": { "type": "integer" },
          "status": { "type": "string" },
          "ingestStage": { "type": "string" },
          "progressPercent": { "type": "integer" },
          "chunkTotal": { "type": "integer", "nullable": true },
          "chunkDone": { "type": "integer" },
          "errorMessage": { "type": "string", "nullable": true },
          "indexedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "IdResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" }
        }
      }
    }
  }
}